Fix python_boreholes., regularize() method
Fixing bug reported in JIRA issue DARDA/DARDA-89. Regularizing data
using the SamplingDomain regularize method removed properties from the
borehole.

Issue occurs because some borehole properties were being incorrectly
designated as non-numeric data so were being skipped. As this is a
dataset-dependent issue (we don't know what data are numeric ahead of
time) this is the right behavior, but perhaps there needs to be
something in the documentation about defining data properly.

However the skipping was happening silently so datasets appeard to
vanish and this was confusing to users. Fix: added a warning when
properties are being skipped by the regularize method.